feat: separate SetSpellingMode component
